requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
coverage==7.10.4
fakeredis==2.31.0
freezegun==1.5.5
gherkin-official==29.0.0
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
parse==1.20.2
parse_type==0.6.6
pluggy==1.6.0
pybreaker==1.4.0
Pygments==2.19.2
pytest==8.4.1
pytest-bdd==8.1.0
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
six==1.17.0
sortedcontainers==2.4.0
tenacity==9.1.2
typing_extensions==4.14.1
urllib3==2.5.0
